Big Code Clean update
Went through all the notebooks and rewrote them in a much cleaner and more concise way. Adapted the earlier notebook to also work with the handcarry data. Made some of the code much more efficient by only reading in a slice of root trees as numpy library instead of the full array. Added much context to the GalaxyFunctions python file.